cleanup.c OK
