+++ cleanup.c
@@ -14,19 +14,6 @@
 
 // Norm: Too many functions in the file
 
-void	ft_clean_minishell_env(void *env_content_node);
-
-/**
- * Save free the pointer. The NULL exuation need to be up in the function
- * otherwise it won't work.
- *
-*/
-void	free_ptr(void *ptr)
-{
-	if (ptr != NULL)
-		free(ptr);
-}
-
 /* restore_io:
 *	Restores the original standard input and standard output
 *	to their original fds of 0 and 1. Used to clear the input/output
@@ -57,51 +44,6 @@ bool	restore_io(t_io_fds *io)
 	return (ret);
 }
 
-/* free_io:
-*	Frees the input/output fd structure.
-*/
-void	free_io(t_io_fds *io)
-{
-	if (!io)
-		return ;
-	restore_io(io);
-	if (io->heredoc_delimiter)
-	{
-		unlink(io->infile);
-		free_ptr(io->heredoc_delimiter);
-	}
-	if (io->infile)
-		free_ptr(io->infile);
-	if (io->outfile)
-		free_ptr(io->outfile);
-	if (io)
-		free_ptr(io);
-}
-
-/* free_str_tab:
-*	Frees an array of strings.
-*/
-void	free_str_tab(char **tab)
-{
-	int	i;
-
-	i = 0;
-	if (tab)
-	{
-		while (tab[i])
-		{
-			if (tab[i])
-			{
-				free_ptr(tab[i]);
-				tab[i] = NULL;
-			}
-			i++;
-		}
-		free(tab);
-		tab = NULL;
-	}
-}
-
 void	lst_delone_cmd(t_command *lst, void (*del)(void *))
 {
 	if (lst->command)
@@ -126,33 +68,6 @@ void	lst_clear_cmd(t_command **lst, void (*del)(void *))
 	}
 }
 
-/* free_data:
-*	Frees all of the data used to run a command. If clear_history is true,
-*	frees the environment and the command history before returning.
-*/
-void	free_global(t_global *global, bool clear_history)
-{
-	if (global && global->user_input)
-	{
-		free_ptr(global->user_input);
-		global->user_input = NULL;
-	}
-	// if (global && global->token)
-	// 	lstclear_token(&global->token, &free_ptr);
-	if (global && global->cmd)
-		lst_clear_cmd(&global->cmd, &free_ptr);
-	// if (clear_history)
-	// {
-	// 	if (global && global->working_dir)
-	// 		free_ptr(global->working_dir);
-	// 	if (global && global->old_working_dir)
-	// 		free_ptr(global->old_working_dir);
-	// 	if (global && global->env)
-	// 		free_str_tab(global->env);
-	// 	rl_clear_history();
-	// }
-}
-
 /* exit_shell:
 *	Cleanly exits the minishell program by closing all opened
 *	fds and freeing all allocated memory.
@@ -161,9 +76,6 @@ void	exit_shell(t_global *global, int exno)
 {
 	if (global)
 	{
-		// TODO uncomment below once the commands are ready
-		// if (global->cmd && global->cmd->io_fds)
-			// close_fds(global->cmd, true);
 		free_global(global, true);
 		if (global && global->env)
 			ft_lstclear(&global->env, ft_clean_minishell_env);
